pyproject.toml
[build-system]
requires = ["flit_core>=3.2.0,<4"]
build-backend = "flit_core.buildapi"


[project]
name = "cosmospy"
version = "6.0.0"  # DO NOT EDIT THIS LINE MANUALLY. LET bump2version UTILITY DO IT
description = "Tools for Cosmos wallet management and offline transaction signing"
authors = [
    { name = "Taneli Hukkinen", email = "hukkin@users.noreply.github.com" },
]
license = { file = "LICENSE" }
requires-python = ">=3.7"
readme = "README.md"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
keywords = ["cosmos", "blockchain", "atom", "cryptocurrency"]
dependencies = [
    "ecdsa >=0.14.0",
    "bech32 >=1.1.0",
    "mnemonic >=0.19",
    "hdwallets >=0.1.0",
    "typing-extensions >=3.7.4; python_version < '3.8'",
    "requests==2.26.0",
    "protobuf >= 3.19.1",
    "google >= 3.0.0",
    "grpcio >=1.46.3"
]

[project.urls]
"Homepage" = "https://github.com/hukkin/cosmospy"
"Changelog" = "https://github.com/hukkin/cosmospy/blob/master/CHANGELOG.md"


[tool.black]
line-length = 99


[tool.isort]
# Force imports to be sorted by module, independent of import type
force_sort_within_sections = true
# Group first party and local folder imports together
no_lines_before = ["LOCALFOLDER"]

# Configure isort to work without access to site-packages
known_first_party = ["cosmospy", "tests"]

# Settings for Black compatibility
profile = "black"
line_length = 99


[tool.ward]
path = ["tests"]
order = "random"


[tool.coverage.run]
source = ["cosmospy"]


[tool.tox]
legacy_tox_ini = '''
[tox]
isolated_build = True

[testenv:py{37,38,39,310}]
description = run tests
deps = -r tests/requirements.txt
commands =
    ward {posargs}
'''


[tool.mypy]
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
# Disabling incremental mode is required for `warn_unused_configs = true` to work
incremental = false
disallow_untyped_defs = true
check_untyped_defs = true
strict_equality = true
implicit_reexport = false
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "ecdsa.*"
ignore_missing_imports = true
